Refactored trigger.py
